supported PPC Mac. Removed error when Stop Speaking is clicked after speech is over.

+++ chrome/content/speakingfox.js
@@ -150,8 +150,10 @@ SpeakingFox.prototype = {
 
   stop: function() {
     try {
-      this.speaking = false;
-      this.processComponent.kill();
+      if(this.speaking) {
+        this.speaking = false;
+        this.processComponent.kill();
+      }
     } catch (err) {
       alert(err);
     }
